Rearranged config files
